requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
requests==2.31.0
google-generativeai==0.3.2
openai==1.3.5
python-dotenv==1.0.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
eventlet==0.33.3
uuid==1.30
PyYAML==6.0.1